Comment out shutdown on CAN error code.
This is an interesting way to initiate safe shutdown of the pi.
With a delayed shut down timer, once CAN messages stop coming
becasue the car has been turned off, the pi can be safely
turned off. But when testing it can be very confusing as you
likely wont be sending lots of can messages and they will
liekly be bad, lol.

Probably rollback this commit once the setup is installed in
the car.