Add openInverterDecoder.js
This is just a slightly modified copy of racePackDecoder.js
Adds support for 11 bit can Ids and overloads BATT_VOLTAGE
with an 11 bit can Idin order to get data from open inverter.

The openInverter as a new decoder part is untested. Testing
was preformed by just changing racePackDecoder.js, but this
seems cleaner.

Removes the ability to specifiy an400. Need to actually remove
the an400 code and likely the racepack stuff too.